more on types and casts
git-svn-id: https://svn.r-project.org/R/trunk@59096 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/appl/stem.c
@@ -114,9 +114,9 @@ stem_leaf(double *x, int n, double scale, int width, double atom)
     i = 0;
     do {
 	if(lo < 0)
-	    stem_print(hi,lo,ndigits);
+	    stem_print((int)hi, (int)lo, ndigits);
 	else
-	    stem_print(lo,hi,ndigits);
+	    stem_print((int)lo, (int)hi, ndigits);
 	j = 0;
 	do {
 	    if(x[i] < 0)xi = (int) (x[i]*c - .5);
