+++ src/main/par-common.c
@@ -54,22 +54,22 @@
     else if (streql(what, "cex.main")) {
 	lengthCheck(what, value, 1, call);	x = asReal(value);
 	posRealCheck(x, what);
-	R_DEV__(cexmain) = x;
+	R_DEV__(cexmain) = (float) x;
     }
     else if (streql(what, "cex.lab")) {
 	lengthCheck(what, value, 1, call);	x = asReal(value);
 	posRealCheck(x, what);
-	R_DEV__(cexlab) = x;
+	R_DEV__(cexlab) = (float) x;
     }
     else if (streql(what, "cex.sub")) {
 	lengthCheck(what, value, 1, call);	x = asReal(value);
 	posRealCheck(x, what);
-	R_DEV__(cexsub) = x;
+	R_DEV__(cexsub) = (float) x;
     }
     else if (streql(what, "cex.axis")) {
 	lengthCheck(what, value, 1, call);	x = asReal(value);
 	posRealCheck(x, what);
-	R_DEV__(cexaxis) = x;
+	R_DEV__(cexaxis) = (float) x;
     }
     else if (streql(what, "col")) {
 	/* col can be a vector of length > 1, so pick off first value
@@ -206,9 +206,10 @@
 	R_DEV__(pch) = ix;
     }
     else if (streql(what, "smo")) {
+	/* FIXME: not real */
 	lengthCheck(what, value, 1, call);	x = asReal(value);
 	nonnegRealCheck(x, what);
-	R_DEV__(smo) = x;
+	R_DEV__(smo) = (int) x;
     }
     else if (streql(what, "srt")) {
 	lengthCheck(what, value, 1, call);	x = asReal(value);
