+++ src/main/par.c
@@ -273,7 +273,7 @@ static void Specify(const char *what, SEXP value, pGEDevDesc dd, SEXP call)
 	lengthCheck(what, value, 1, call);	x = asReal(value);
 	posRealCheck(x, what);
 	R_DEV__(cex) = 1.0; /* ! (highlevel par, i.e.  Specify2(), set x ! */
-	R_DEV__(cexbase) = x;
+	R_DEV__(cexbase) = (float) x;
     }
 
     else if (streql(what, "fg")) {
@@ -405,11 +405,11 @@ static void Specify(const char *what, SEXP value, pGEDevDesc dd, SEXP call)
 	R_DEV_2(defaultFigure) = TRUE;
 	R_DEV_2(layout) = FALSE;
 	if (nrow > 2 || ncol > 2) {
-	    R_DEV_2(cexbase) = 0.66;
+	    R_DEV_2(cexbase) = (float) 0.66;
 	    R_DEV_2(mex) = 1.0;
 	}
 	else if (nrow == 2 && ncol == 2) {
-	    R_DEV_2(cexbase) = 0.83;
+	    R_DEV_2(cexbase) = (float) 0.83;
 	    R_DEV_2(mex) = 1.0;
 	}
 	else {
@@ -434,11 +434,11 @@ static void Specify(const char *what, SEXP value, pGEDevDesc dd, SEXP call)
 	R_DEV_2(defaultFigure) = TRUE;
 	R_DEV_2(layout) = FALSE;
 	if (nrow > 2 || ncol > 2) {
-	    R_DEV_2(cexbase) = 0.66;
+	    R_DEV_2(cexbase) = (float) 0.66;
 	    R_DEV_2(mex) = 1.0;
 	}
 	else if (nrow == 2 && ncol == 2) {
-	    R_DEV_2(cexbase) = 0.83;
+	    R_DEV_2(cexbase) = (float) 0.83;
 	    R_DEV_2(mex) = 1.0;
 	}
 	else {
@@ -1263,11 +1263,11 @@ SEXP attribute_hidden do_layout(SEXP call, SEXP op, SEXP args, SEXP env)
     /*------------------------------------------------------*/
 
     if (nrow > 2 || ncol > 2) {
-	gpptr(dd)->cexbase = dpptr(dd)->cexbase = 0.66;
+	gpptr(dd)->cexbase = dpptr(dd)->cexbase = (float) 0.66;
 	gpptr(dd)->mex = dpptr(dd)->mex = 1.0;
     }
     else if (nrow == 2 && ncol == 2) {
-	gpptr(dd)->cexbase = dpptr(dd)->cexbase = 0.83;
+	gpptr(dd)->cexbase = dpptr(dd)->cexbase = (float) 0.83;
 	gpptr(dd)->mex = dpptr(dd)->mex = 1.0;
     }
     else {
